moves JS code to it's own file for signup
